src/monolithic/requirements.txt
fastapi[all]==0.89.1
psycopg2==2.9.5
sqlalchemy==1.4.46
sqlalchemy-cockroachdb==1.4.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
gunicorn==20.1.0
